style: 코드리뷰 반영 및 Lint 수정
